SIMULATOR: Fixed velocity and gps time output.
The simulator previously did not send gps_time or vel_ned when simulating an RTK-style fix. We now always send gps_time and vel_ned.

The simulator will now also experience a gps_time week rollover if started 20 seconds after boot. This is useful for testing drivers. We do not yet support simulating arbitrary gps_time, since we simulate based on a hard-coded almanac that is only valid for a ~4h window.