[work in progress] handle acq to track handover (#947)
Looks like HITL results are close enough to master, and this seems a better way to handle acq to tracking handover as it does not do priority inversion.. more failures to serve NAP interrupts might be due to removing the deadline-stretching loop in `nap_common.c` but I think that is more elegant and stretching should not be needed. 
It is important now to see which other places lock the CPU temporarily using `chSysLock()` for example and prevent the `nap_track_irq` from closing the loop.