Update Makefiles to link with libswiftnav.a.
The incoming CMake changes in LSNP remove the libswiftnav-static
library in favor of only building libswiftnav. Whether or not it is
a dynamic object library is controlled by the BUILD_SHARED_LIBS CMake
variable. In PFWP, this is set to OFF by the ARM toolchain file, so
the behavior is the same as before in here.

Note: The location of libfec is also changed.