ndb: Fix bug creating new files.
Failure to increment the offset in `ndb_fs_reserve` meant that
the start of the file is repeatedly written with zeros but
the bulk of the file is left unchanged.  This hasn't been a
problem, as the file is removed first, and when the new file is
created, writing the end record automatically pads the file
with zeros.

This PR also changes `ndb_create_file` to only
reserve the file with this method if the unlink fails.  This
saves a significant amount of traffic clearing the ephemeris
on start up.