starling platform shim 3 (#1870)
* Time-matched obs synchronization into platform.

The buff pool and mailbox handling is now entirely in the platform
layer. The time-matched thread should be more-or-less Chibi agnostic
at this point.

* Replaced base-obs mailbox with platform call.

* Move me_msg_t passing into platform layer.

Also removed the generic platform mailbox and pool functions. This
should allow the other calc_* headers to be removed as soon as the
common types are separated.