Overhaul to Starling settings handling.
All settings are now synchronized and managed from a single
struct. Prior to every filter update, the most recent settings
are grabbed and applied to the relevant filter manager. This means
that settings may be modified from any thread at any time, and they
will be applied to the filter managers for all subsequent update steps.