Use the proper platform thread function.
This required some shuffling of the integration code, and
also a bit of a hack. Now that the low-latency and time-matched
threads are being started simultaneously, there is a little
bit of extra synchronization required to avoid a race condition
around the initialization of the filter managers. At the moment
this handled by a bespoke semaphore in the platform shim. Once
the formal mailbox implementation is complete, the threads
can communicate with mailbox messages properly.