Remove check for FCNs in Glonass map on acquisition
Without this change acquisition will not be attempted on frequencies already
mapped to slots.

One problem is that it will not search sats which fcn are present twice
in the table, and we should do it after some time (11-12H basically)

Another is that it will not search sats which fcn are present once in the
table and will just spin in a loop until the watchdog kicks.