Restore thread priorities to how they were in 2.4
In 2.4, the Starling thread had a priority of `HIGHPRIO-4`. It appears
that it has since been inadvertently changed to `NORMALPRIO-4` since,
which makes it a lower priority thread than the time matched thread.

There hasn't been anything found that has been broken because of this
change, but in the interest of consistency, it has been reverted to what
it previously was.