Fix error handling in fileio read/write
The previously defined BREAK_ON_FAIL macro was defined within a
do-while block, so the break within the macro was only breaking out
of the do-while block rather than the parent while loop. This meant that
fileio operations would continue even after an error occurring, which
leads to crashes and unexpected behaviour when variables haven't been
initialised as expected.

This change removes the macro and puts the log/break directly into the
read/write loops.