imu: Add autorecovery (PIKSI-98)
Adds the ability for the firmware to hard reset the IMU if the data
stream unexpectedly stops. This is to fix an issue found during EMC &
ESD testing where the IMU data stopped when the Duro was zapped. Details
around the exact failure mode that caused this issue are unclear,
however it is hopeful that performing a hard reset in this way would
have helped.

This was tested by adding a setting that would set the IMU_EN signal
low, then checking that the IMU output resumed afterwards.