Add a "ALWAYS" PPS option (#4468)
The always setting makes the PPS signal get emitted from boot, even
before there is fix. When there is no fix, it will be emitted at the
specified frequency, width, etc but will only be synchronized to the
boot time of the Piksi.

The implementation relies on the fact that the persistent_clock_state
time is initialized to Jan 1980 in main.c at boot.

Once there is a fix, the PPS signal will ungracefully 'jump' to be
properly synchronized with the GPS time.