starling - jenkins: Fix Linux Perf CI Stage [BUILD-194] * Fix Linux Perf Stage
Removes the dependency on the source tree during the linux perf
profiling stage. This makes it so that the repo does not need to
be cloned during this stage. The git checkout is a persistent
source of pipeline failures due to network flakiness on the build
agent we use for this job.

* Add error handling

Adds try/finally/catch to the pipeline so that worker nodes get cleaned
up properly between builds.

* Pin Base Image

Changes the `latest` tag to a stable ubuntu base for the image.
Using pinned images avoids potential base image upgrade surprises.

Triggered-By:	starling	066c0aac9adffcab16e3b9fadbe8ae3dd98bf43f
Upstream-PR:	http://github.com/swift-nav/starling/pull/6653